+++ src/wrapped_radix_engine_toolkit/low_level/transaction_classes/detailed_manifest_class.rs
@@ -6,49 +6,92 @@ use radix_engine_toolkit::transaction_types::{
     Operation as RetOperation,
 };
 
-use radix_engine_interface::blueprints::resource::ResourceOrNonFungible as ScryptoResourceOrNonFungible;
+/// The execution summary process not only determines the class of the manifest,
+/// but also includes additional information about this class that the wallet
+/// requires to display to the user.
+#[derive(Clone, Debug, PartialEq, Eq, EnumAsInner, uniffi::Enum)]
+pub enum DetailedManifestClass {
+    /// A general manifest that involves any amount of arbitrary components
+    /// and packages where nothing more concrete can be said about the manifest
+    /// and its nature.
+    ///
+    /// No additional information is required beyond what the execution summary
+    /// will provide.
+    General,
 
-pub(crate) fn to_vec_network_aware<T, U>(
-    values: impl IntoIterator<Item = T>,
-    network_id: NetworkID,
-) -> Vec<U>
-where
-    U: From<(T, NetworkID)>,
-{
-    values
-        .into_iter()
-        .map(|x| (x, network_id))
-        .map(U::from)
-        .collect_vec()
-}
+    /// A manifest of a 1-to-1 transfer to a one-to-many transfer of resources.
+    Transfer {
+        /// When `true`, then this is a one-to-one transfer and the wallet can
+        /// regard this as a "simple transfer" and communicate this information
+        /// to the ledger hardware wallet. Otherwise, if `false`, then this is
+        /// not a one-to-one transfer.
+        is_one_to_one: bool,
+    },
 
-pub(crate) fn to_hashmap_network_aware_key<K, V, L, U>(
-    values: impl IntoIterator<Item = (K, V)>,
-    network_id: NetworkID,
-) -> HashMap<L, U>
-where
-    L: Eq + std::hash::Hash + From<(K, NetworkID)>,
-    U: From<V>,
-{
-    values
-        .into_iter()
-        .map(|(k, v)| (L::from((k, network_id)), U::from(v)))
-        .collect::<HashMap<L, U>>()
-}
+    /// A manifest where XRD is claimed from one or more validators.
+    ValidatorClaim {
+        /// The addresses of validators in the transaction
+        validator_addresses: Vec<ValidatorAddress>,
+        /// The claims observed in the transaction
+        validator_claims: Vec<TrackedValidatorClaim>,
+    },
 
-pub(crate) fn filter_try_to_vec_network_aware<T, U>(
-    values: impl IntoIterator<Item = T>,
-    network_id: NetworkID,
-) -> Vec<U>
-where
-    U: TryFrom<(T, NetworkID)>,
-{
-    values
-        .into_iter()
-        .map(|x| (x, network_id))
-        .map(U::try_from)
-        .filter_map(Result::ok)
-        .collect_vec()
+    /// A manifest where XRD is staked to one or more validators.
+    ValidatorStake {
+        /// The addresses of validators in the transaction
+        validator_addresses: Vec<ValidatorAddress>,
+        /// The stake observed in the transaction
+        validator_stakes: Vec<TrackedValidatorStake>,
+    },
+
+    /// A manifest where XRD is unstaked from one or more validators.
+    ValidatorUnstake {
+        /// The addresses of validators in the transaction
+        validator_addresses: Vec<ValidatorAddress>,
+
+        /// The data associated with the various claim NFTs
+        claims_non_fungible_data: HashMap<NonFungibleGlobalId, UnstakeData>,
+    },
+
+    /// A manifest that updated the deposit settings of the account.
+    AccountDepositSettingsUpdate {
+        /// Updates to the resource preferences of the account deposit settings.
+        /// account_address -> (resource_address -> Update<new_preference>)
+        resource_preferences_updates: HashMap<
+            AccountAddress,
+            HashMap<ResourceAddress, ResourcePreferenceUpdate>,
+        >,
+        /// Changes to the account's deposit mode.
+        /// account_address -> new_default_deposit_mode
+        deposit_mode_updates: HashMap<AccountAddress, DepositRule>,
+        /// Additions to the authorized depositors
+        authorized_depositors_added:
+            HashMap<AccountAddress, Vec<ResourceOrNonFungible>>,
+        /// Removals from the authorized depositors
+        authorized_depositors_removed:
+            HashMap<AccountAddress, Vec<ResourceOrNonFungible>>,
+    },
+
+    /// A manifest that contributed some amount of resources to a liquidity
+    /// pool that can be a one-resource pool, two-resource pool, or a
+    /// multi-resource pool.
+    PoolContribution {
+        /// The addresses of the pools in the transaction
+        pool_addresses: Vec<PoolAddress>,
+        /// The contribution observed in the transaction
+        pool_contributions: Vec<TrackedPoolContribution>,
+    },
+
+    /// A manifest that redeemed resources from a liquidity pool. Similar to
+    /// contributions, this can be any of the three pool blueprints available
+    /// in the pool package.
+    PoolRedemption {
+        /// The addresses of the pools in the transaction
+        pool_addresses: Vec<PoolAddress>,
+
+        /// The redemptions observed in the transaction
+        pool_contributions: Vec<TrackedPoolRedemption>,
+    },
 }
 
 impl From<(RetDetailedManifestClass, NetworkID)> for DetailedManifestClass {
@@ -57,8 +100,8 @@ impl From<(RetDetailedManifestClass, NetworkID)> for DetailedManifestClass {
         match value.0 {
             RetDetailedManifestClass::General => Self::General,
 
-            RetDetailedManifestClass::Transfer { is_one_to_one: _ } => {
-                Self::Transfer
+            RetDetailedManifestClass::Transfer { is_one_to_one } => {
+                Self::Transfer { is_one_to_one }
             }
 
             RetDetailedManifestClass::PoolContribution {
@@ -182,63 +225,4 @@ impl From<(RetDetailedManifestClass, NetworkID)> for DetailedManifestClass {
             }
         }
     }
-}
-
-impl From<(ScryptoResourceOrNonFungible, NetworkID)> for ResourceOrNonFungible {
-    fn from(value: (ScryptoResourceOrNonFungible, NetworkID)) -> Self {
-        let (resource_or_non_fungible, network_id) = value;
-        match resource_or_non_fungible {
-            ScryptoResourceOrNonFungible::NonFungible(nf) => {
-                Self::NonFungible {
-                    value: (nf, network_id).into(),
-                }
-            }
-            ScryptoResourceOrNonFungible::Resource(resource_address) => {
-                Self::Resource {
-                    value: (resource_address, network_id).into(),
-                }
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, EnumAsInner, uniffi::Enum)]
-pub enum DetailedManifestClass {
-    General,
-    Transfer,
-
-    ValidatorClaim {
-        validator_addresses: Vec<ValidatorAddress>,
-        validator_claims: Vec<TrackedValidatorClaim>,
-    },
-
-    ValidatorStake {
-        validator_addresses: Vec<ValidatorAddress>,
-        validator_stakes: Vec<TrackedValidatorStake>,
-    },
-
-    ValidatorUnstake {
-        validator_addresses: Vec<ValidatorAddress>,
-        claims_non_fungible_data: HashMap<NonFungibleGlobalId, UnstakeData>,
-    },
-
-    AccountDepositSettingsUpdate {
-        resource_preferences_updates: HashMap<
-            AccountAddress,
-            HashMap<ResourceAddress, ResourcePreferenceUpdate>,
-        >,
-        deposit_mode_updates: HashMap<AccountAddress, DepositRule>,
-        authorized_depositors_added:
-            HashMap<AccountAddress, Vec<ResourceOrNonFungible>>,
-        authorized_depositors_removed:
-            HashMap<AccountAddress, Vec<ResourceOrNonFungible>>,
-    },
-    PoolContribution {
-        pool_addresses: Vec<PoolAddress>,
-        pool_contributions: Vec<TrackedPoolContribution>,
-    },
-    PoolRedemption {
-        pool_addresses: Vec<PoolAddress>,
-        pool_contributions: Vec<TrackedPoolRedemption>,
-    },
 }
