+++ src/wrapped_radix_engine_toolkit/low_level/transaction_classes/types/mod.rs
@@ -1,11 +1,11 @@
-mod resource_preferences_update;
+mod resource_preference_update;
 mod tracked_pool_contribution;
 mod tracked_pool_redemption;
 mod tracked_validator_claim;
 mod tracked_validator_stake;
 mod unstake_data;
 
-pub use resource_preferences_update::*;
+pub use resource_preference_update::*;
 pub use tracked_pool_contribution::*;
 pub use tracked_pool_redemption::*;
 pub use tracked_validator_claim::*;
