+++ src/wrapped_radix_engine_toolkit/low_level/transaction_classes/types/resource_preference_update.rs
@@ -1,3 +1,5 @@
+use crate::prelude::*;
+
 use radix_engine_interface::blueprints::account::ResourcePreference as RetResourcePreference;
 use radix_engine_toolkit::transaction_types::Update as RetUpdate;
 
@@ -30,5 +32,51 @@ impl From<RetUpdate<RetResourcePreference>> for ResourcePreferenceUpdate {
             },
             RetUpdate::Remove => Self::Remove,
         }
+    }
+}
+
+impl HasSampleValues for ResourcePreferenceUpdate {
+    fn sample() -> Self {
+        Self::Set {
+            value: ResourcePreference::Allowed,
+        }
+    }
+
+    fn sample_other() -> Self {
+        Self::Remove
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[allow(clippy::upper_case_acronyms)]
+    type SUT = ResourcePreferenceUpdate;
+
+    #[test]
+    fn equality() {
+        assert_eq!(SUT::sample(), SUT::sample());
+        assert_eq!(SUT::sample_other(), SUT::sample_other());
+    }
+
+    #[test]
+    fn inequality() {
+        assert_ne!(SUT::sample(), SUT::sample_other());
+    }
+
+    #[test]
+    fn from_ret() {
+        assert_eq!(
+            SUT::from(RetUpdate::<RetResourcePreference>::Set(
+                RetResourcePreference::Allowed
+            )),
+            SUT::sample()
+        );
+
+        assert_eq!(
+            SUT::from(RetUpdate::<RetResourcePreference>::Remove),
+            SUT::sample_other()
+        );
     }
 }
